requirements.txt
requests
beautifulsoup4
pandas
psycopg2-binary


